feat: resolve dynamic rating engine and implement full favorites workflow

+++ CartFlow.Web/Models/ProductViewModel.cs
@@ -8,26 +8,19 @@
         public decimal UnitPrice { get; set; }
         public int StockQuantity { get; set; }
 
-        // Denormalised from the related Category entity
         public string? CategoryName { get; set; }
         public int CategoryId { get; set; }
 
-        // URL of the first ProductImage (null when no images exist)
         public string? ImageUrl { get; set; }
-
-        // All product image URLs (used by the details gallery)
         public List<string> ImageUrls { get; set; } = new();
 
-        // First letter of the product name — used as a fallback avatar/placeholder in the UI
         public string Initial { get; set; } = string.Empty;
 
-        // Reviews (populated from IReviewService). Uses DTOs from Services layer to avoid circular deps.
+        // القائمة الحقيقية للمراجعات
         public List<CartFlow.Services.Models.ReviewDto> Reviews { get; set; } = new();
 
-        // Convenience properties for the index view
-        public decimal AverageRating => Reviews?.Any() == true ? Math.Round(Reviews.Average(r => r.Rate), 2) : 0m;
+        // الخصائص الذكية المحسوبة ديناميكياً تلقائياً بمجرد ملء القائمة أعلاه 🔥
+        public decimal AverageRating => Reviews?.Any() == true ? (decimal)Math.Round(Reviews.Average(r => r.Rate), 2) : 0m;
         public int ReviewCount => Reviews?.Count() ?? 0;
-
-        // Calculated average rating based on Reviews
     }
 }
